% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2pipes.R
\name{create_pipe_enabled_ggplot2_func}
\alias{create_pipe_enabled_ggplot2_func}
\title{Create a pipe-aware func from any plus-aware function}
\usage{
create_pipe_enabled_ggplot2_func(ggplot2_func_name, prefix = "add_")
}
\arguments{
\item{ggplot2_func_name}{Name of a ggplot2 function. character. e.g. "geom_point"}

\item{prefix}{string with which to prefix names of \code{ggplot2} functions in order to name the pipe-enabled functions. default: "add_".
Note: You could set this the empty string, in which case the new functions would mask the name of the library function}
}
\description{
Assigns the new function in the global environemnt
}
