% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2pipes.R
\name{init_ggplot2_pipes}
\alias{init_ggplot2_pipes}
\title{augment a packages which uses composition via '+' into package which understand pipes}
\usage{
init_ggplot2_pipes(prefix = "add_")
}
\arguments{
\item{prefix}{string with which to prefix names of \code{ggplot2} functions in order to name the pipe-enabled functions}
}
\description{
augment a packages which uses composition via '+' into package which understand pipes
}
\details{
This is proof-of-concept only, which was inspired after seeing @hadleywickham
and @cpsievert discussing DSLs for plotting:
https://twitter.com/cpsievert/status/606190120568365056
}
